WIP: Cross-compile QT packages from native chroot
I'm starting with qt5-qtwebchannel, because that is a small package. The
goal is to compile qt5-qtwebengine eventually (the whole chromium
browser).

In the current status, this does not work, because we can't run
configure in the QT module's source (it is only shipped in the main
source tree, not in the module's trees). So right now it compiles code
with the native compiler instead of the cross-compiler, and fails at
stripping the binaries with the foreign-arch strip command (abuild runs
it).

TODO:
- Find a way to tell qmake to cross-compile
- Make it work with qt5-qtwebchannel
- Make it work with qt5-qtwebengine
- Split up init_builenv into two functions?
- Use shlex to escape the environment variables we set for cross-
  compiling instead of "\"...\""

Evnironment variables code from:
<https://github.com/postmarketOS/pmbootstrap/pull/665>