Make /home/pmos/build world-readable
When calculating checksums, we copy an aport folder to
`/home/pmos/build`, `chown` it to the `pmos` user, let `abuild`
calculate the checksums, and copy the `APKINDEX` back.

So far this only worked as long as the `aports/main/hello-world`
folder in this example is already world-readable. If it is not (e.g.
because the system had a different default umask), then the copying
back step will fail.

This commit makes the copied folder world readable to fix this.