pmb: give loop module time to initialize
During `pmbootstrap install`, we need the `loop` kernel module in
order to write to the rootfs image file as loopback device.

We already load `loop` on demand. However, during the move of the CI
infrastructure from Travis to gitlab CI, we have noticed that on some
Linux installations, the loop devices are not available instantly.

With this commit, pmbootstrap tries to mount the loopback device up
to five times, waiting one second between each attempt.