Wiki CI: require all devices to be booting
In gitlab, there is no extra CI job running for commits in a merge
request (MR). This means, we can't run different code in a MR against
"master".

So instead of only checking whether all devices are booting if there's
a MR against master, always perform this check.

I've edited the message slightly, so it's clear that it's only required
to have the devices booting when making a merge request against master.