deviceinfo: prepend manufacturer to device name
Overview:
About half the devices had the manufacturer name in the
"deviceinfo_name" variable, the other half did not. For consistency in
our aports, and to make reviews consistent as well, it is now mandatory
to prepend the vendor name to the device's name.

I thought about forcing it the other way as well, but then we have a
few weird cases, where the phone has the same name as the company,
"Jolla Phone" would just have `deviceinfo_name="Phone"` then.

Details:
* Add test that requires the manufacturer's name to be in front of
  `deviceinfo_name`
* Adjust all existing `deviceinfo` files which failed this test, and
  bump the pkgrel
* New device wizard: when the manufacturer is not specified in front of
  the name, add it automatically.
* Adjust test case for the `ask_for_name()` function.